+++ geotrek/api/v2/mixins.py
@@ -7,7 +7,7 @@ from geotrek.common import models as common_models
 
 class PDFSerializerMixin:
 
-    def _get_pdf_url_lang(self, obj, lang):
+    def _get_pdf_url_lang(self, obj, lang, portal=None):
         namespace = self.Meta.model._meta.app_label
         modelname = self.Meta.model._meta.object_name.lower()
         if settings.ONLY_EXTERNAL_PUBLIC_PDF:
@@ -18,15 +18,17 @@ class PDFSerializerMixin:
         url = reverse(urlname, kwargs={'lang': lang, 'pk': obj.pk, 'slug': obj.slug})
         request = self.context.get('request')
         if request:
-            url = request.build_absolute_uri(url)
+            url = f'{request.build_absolute_uri(url)}?portal={portal.split(",")[0]}' \
+                if portal else request.build_absolute_uri(url)
         return url
 
     def get_pdf_url(self, obj):
         lang = self.context.get('request').GET.get('language', 'all') if self.context.get('request') else 'all'
+        portal = self.context.get('request').GET.get('portals', None) if self.context.get('request') else None
         if lang != 'all':
-            data = self._get_pdf_url_lang(obj, lang)
+            data = self._get_pdf_url_lang(obj, lang, portal)
         else:
             data = {}
             for language in settings.MODELTRANSLATION_LANGUAGES:
-                data[language] = self._get_pdf_url_lang(obj, language)
+                data[language] = self._get_pdf_url_lang(obj, language, portal)
         return data